More refactoring of tests
