--- tests/Sqlite/DataFilterTest.php
@@ -1,10 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace Yiisoft\Data\Db\Tests\Sqlite;
-
-final class DataFilterTest extends \Yiisoft\Data\Db\Tests\Base\DataFilterTest
-{
-    use DatabaseTrait;
-}
