--- tests/Sqlite/QueryWithFiltersTest.php
@@ -1,10 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace Yiisoft\Data\Db\Tests\Sqlite;
-
-final class QueryWithFiltersTest extends \Yiisoft\Data\Db\Tests\Base\QueryWithFiltersTest
-{
-    use DatabaseTrait;
-}
